refactor: SearchHistory 패키지 분리
