+++ src/main/java/com/techfork/domain/activity/readhistory/repository/SearchHistoryRepository.java
@@ -1,0 +1,15 @@
+package com.techfork.domain.activity.readhistory.repository;
+
+import com.techfork.domain.activity.readhistory.entity.SearchHistory;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+
+import java.util.List;
+
+public interface SearchHistoryRepository extends JpaRepository<SearchHistory, Long> {
+
+    @Query("SELECT sh FROM SearchHistory sh WHERE sh.user.id = :userId ORDER BY sh.searchedAt DESC")
+    List<SearchHistory> findRecentSearchHistoriesByUserId(@Param("userId") Long userId, Pageable pageable);
+}
