+++ src/test/java/com/techfork/domain/activity/readhistory/service/ReadHistoryCommandServiceTest.java
@@ -1,8 +1,8 @@
-package com.techfork.domain.activity.service;
+package com.techfork.domain.activity.readhistory.service;
 
-import com.techfork.domain.activity.dto.SearchHistoryRequest;
-import com.techfork.domain.activity.entity.SearchHistory;
-import com.techfork.domain.activity.repository.SearchHistoryRepository;
+import com.techfork.domain.activity.readhistory.dto.SearchHistoryRequest;
+import com.techfork.domain.activity.readhistory.entity.SearchHistory;
+import com.techfork.domain.activity.readhistory.repository.SearchHistoryRepository;
 import com.techfork.domain.useraccount.entity.User;
 import com.techfork.domain.useraccount.exception.UserErrorCode;
 import com.techfork.domain.useraccount.repository.UserRepository;
@@ -28,7 +28,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-class ActivityCommandServiceTest {
+class ReadHistoryCommandServiceTest {
 
     @Mock
     private UserRepository userRepository;
@@ -37,7 +37,7 @@ class ActivityCommandServiceTest {
     private SearchHistoryRepository searchHistoryRepository;
 
     @InjectMocks
-    private ActivityCommandService activityCommandService;
+    private ReadHistoryCommandService readHistoryCommandService;
 
     @Test
     @DisplayName("검색 히스토리 저장 성공")
@@ -51,7 +51,7 @@ class ActivityCommandServiceTest {
         given(userRepository.findById(userId)).willReturn(Optional.of(mockUser));
         given(searchHistoryRepository.save(any(SearchHistory.class))).willReturn(mock(SearchHistory.class));
 
-        activityCommandService.saveSearchHistory(userId, request);
+        readHistoryCommandService.saveSearchHistory(userId, request);
 
         verify(userRepository, times(1)).findById(userId);
         ArgumentCaptor<SearchHistory> searchHistoryCaptor = ArgumentCaptor.forClass(SearchHistory.class);
@@ -67,7 +67,7 @@ class ActivityCommandServiceTest {
 
         given(userRepository.findById(userId)).willReturn(Optional.empty());
 
-        assertThatThrownBy(() -> activityCommandService.saveSearchHistory(userId, request))
+        assertThatThrownBy(() -> readHistoryCommandService.saveSearchHistory(userId, request))
                 .isInstanceOf(GeneralException.class)
                 .hasFieldOrPropertyWithValue("code", UserErrorCode.USER_NOT_FOUND);
