+++ src/test/java/com/techfork/domain/useraccount/repository/UserRepositoryTest.java
@@ -2,10 +2,10 @@ package com.techfork.domain.useraccount.repository;
 
 import com.techfork.domain.activity.readpost.entity.ReadPost;
 import com.techfork.domain.activity.bookmark.entity.Bookmark;
-import com.techfork.domain.activity.entity.SearchHistory;
+import com.techfork.domain.activity.readhistory.entity.SearchHistory;
 import com.techfork.domain.activity.readpost.repository.ReadPostRepository;
 import com.techfork.domain.activity.bookmark.repository.BookmarkRepository;
-import com.techfork.domain.activity.repository.SearchHistoryRepository;
+import com.techfork.domain.activity.readhistory.repository.SearchHistoryRepository;
 import com.techfork.domain.post.entity.Post;
 import com.techfork.domain.post.repository.PostRepository;
 import com.techfork.domain.source.entity.TechBlog;
